+++ src/labkey/Utils.ts
@@ -340,7 +340,7 @@ export function encode(data: any): string {
  * by the browser. For example, if your input string was "&lt;p&gt;Hello&lt;/p&gt;" the output would be
  * "&amp;lt;p&amp;gt;Hello&amp;lt;/p&amp;gt;". If you set an element's innerHTML property
  * to this string, the HTML markup will be displayed as literal text rather than being
- * interpreted as HTML. By default this function will return an empty string if a value
+ * interpreted as HTML. By default, this function will return an empty string if a value
  * of undefined or null is passed it. To prevent this default, you can pass in a second
  * optional parameter value of true to retain the empty value's type.
  *
@@ -595,13 +595,12 @@ export function getMsgFromError(response: XMLHttpRequest, exceptionObj: any, con
 
 /**
  *
- * Standard documented name for error callback arguments is "failure" but various other names have been employed in past.
+ * Standard documented name for error callback arguments is "failure" but various other names have been employed in the past.
  * This function provides reverse compatibility by picking the failure callback argument out of a config object
  * be it named failure, failureCallback or errorCallback.
  */
 export function getOnFailure(config: { errorCallback?: any; failure?: any; failureCallback?: any }): any {
     return config.failure || config.errorCallback || config.failureCallback;
-    // maybe it be desirable for this fall all the way back to returning LABKEY.Utils.displayAjaxErrorResponse?
 }
 
 /**
@@ -636,10 +635,11 @@ export function id(prefix?: string): string {
     return ID_PREFIX + ++idSeed;
 }
 
-export function isArray(value: any): boolean {
-    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/isArray
-    return Object.prototype.toString.call(value) === '[object Array]';
-}
+/**
+ * Returns true if the passed value is an Array.
+ * @deprecated Use `Array.isArray()`
+ */
+export const isArray = Array.isArray;
 
 /**
  * Tests whether the passed value can be used as boolean, using a loose definition.
@@ -649,40 +649,33 @@ export function isArray(value: any): boolean {
  * @param value The value to test
  * @returns {boolean}
  */
-export function isBoolean(value: any): boolean {
-    const upperVal = value.toString().toUpperCase();
-    if (
-        upperVal == 'TRUE' ||
-        value == '1' ||
-        upperVal == 'Y' ||
-        upperVal == 'YES' ||
-        upperVal == 'ON' ||
-        upperVal == 'T' ||
-        upperVal == 'FALSE' ||
-        value == '0' ||
-        upperVal == 'N' ||
-        upperVal == 'NO' ||
-        upperVal == 'OFF' ||
-        upperVal == 'F'
-    ) {
-        return true;
-    }
+export function isBoolean(value: unknown): value is boolean {
+    if (typeof value === 'boolean') return true;
+    const normalized = value?.toString().toUpperCase();
+    const booleanStrings = new Set(['TRUE', 'FALSE', '1', '0', 'Y', 'N', 'YES', 'NO', 'ON', 'OFF', 'T', 'F']);
+    return booleanStrings.has(normalized);
 }
 
-export function isDate(value: any): boolean {
+/**
+ * Returns true if the passed value is a Date.
+ */
+export function isDate(value: unknown): value is Date {
     return Object.prototype.toString.call(value) === '[object Date]';
 }
 
-export function isNumber(value: any): boolean {
-    return typeof value === 'number' && Number.isFinite(value);
-}
+/**
+ * Returns true if the passed value is a finite number.
+ * @deprecated Use `Number.isFinite()`.
+ * @param value
+ */
+export const isNumber = Number.isFinite;
 
 export function isDefined(value: any): boolean {
     return typeof value !== 'undefined';
 }
 
 /**
- * Returns true if the passed object is empty (ie. `{}`) and false if not.
+ * Returns true if the passed object is empty (i.e. `{}`) and false if not.
  * @param obj The object to test
  * @return The result of the test
  */